Guest checkout

+++ client/store/currentOrder.js
@@ -22,6 +22,15 @@ export const fetchOrder = () => {
           },
         });
         dispatch(setCurrentOrder(data));
+      } else {
+        const cart = JSON.parse(localStorage.getItem("cart"))
+        if (!cart) {
+          let cart = {
+            products: []
+          }
+          localStorage.setItem("cart", JSON.stringify(cart))
+        }
+        dispatch(setCurrentOrder(cart))
       }
     } catch (err) {
       console.error(err);
@@ -33,15 +42,46 @@ export const addToOrder = (productIdQuantity) => {
   return async (dispatch) => {
     try {
       const token = window.localStorage.getItem(TOKEN);
-      const addedProd = await axios.post(
-        '/api/orders/current',
-        productIdQuantity,
-        {
-          headers: {
-            authorization: token,
-          },
+      if (token) {
+        const addedProd = await axios.post(
+          '/api/orders/current',
+          productIdQuantity,
+          {
+            headers: {
+              authorization: token,
+            },
+          }
+        );
+      } else {
+        let cart = JSON.parse(localStorage.getItem("cart"))
+        if (cart) {
+
+          let x = true;
+          cart.products.forEach(element => {
+            if (element.id === productIdQuantity.product.id) {
+              element.orderproduct.quantity += productIdQuantity.quantity
+              x = false
+            }
+          });
+
+          if (x) {
+            cart.products.push({
+              ...productIdQuantity.product,
+              orderproduct: { quantity: productIdQuantity.quantity }
+            })
+          }
+
+          localStorage.setItem("cart", JSON.stringify(cart))
+        } else {
+          cart = {
+            products: [{
+              ...productIdQuantity.product,
+              orderproduct: { quantity: productIdQuantity.quantity }
+            }]
+          }
+          localStorage.setItem("cart", JSON.stringify(cart))
         }
-      );
+      }
     } catch (err) {
       console.error(err);
     }
@@ -52,17 +92,29 @@ export const updateOrder = (orderDetails) => {
   return async (dispatch) => {
     try {
       const token = window.localStorage.getItem(TOKEN);
-      await axios.put('/api/orders/current', orderDetails, {
-        headers: {
-          authorization: token,
-        },
-      });
-      const { data: newOrder } = await axios.get('/api/orders/current', {
-        headers: {
-          authorization: token,
-        },
-      });
-      dispatch(setCurrentOrder(newOrder));
+      if (token) {
+        await axios.put('/api/orders/current', orderDetails, {
+          headers: {
+            authorization: token,
+          },
+        });
+        const { data: newOrder } = await axios.get('/api/orders/current', {
+          headers: {
+            authorization: token,
+          },
+        });
+        dispatch(setCurrentOrder(newOrder));
+      } else {
+        let cart = JSON.parse(localStorage.getItem("cart"))
+
+        cart.products.forEach(element => {
+          if (orderDetails.updated[element.id]) {
+            element.orderproduct.quantity = orderDetails.updated[element.id]
+          }
+        });
+        localStorage.setItem("cart", JSON.stringify(cart))
+        dispatch(setCurrentOrder(cart));
+      }
     } catch (err) {
       console.error(err);
     }
@@ -72,17 +124,25 @@ export const confirmOrder = (orderId) => {
   return async (dispatch) => {
     try {
       const token = window.localStorage.getItem(TOKEN);
-      await axios.put('/api/orders/current?confirmed=true', orderId, {
-        headers: {
-          authorization: token,
-        },
-      });
-      const { data: newOrder } = await axios.get('/api/orders/current', {
-        headers: {
-          authorization: token,
-        },
-      });
-      dispatch(setCurrentOrder(newOrder));
+      if (token) {
+        await axios.put('/api/orders/current?confirmed=true', orderId, {
+          headers: {
+            authorization: token,
+          },
+        });
+        const { data: newOrder } = await axios.get('/api/orders/current', {
+          headers: {
+            authorization: token,
+          },
+        });
+        dispatch(setCurrentOrder(newOrder));
+      } else {
+        localStorage.removeItem("cart")
+        let cart = {
+          products: []
+        }
+        localStorage.setItem("cart", JSON.stringify(cart))
+      }
     } catch (err) {
       console.error(err);
     }
@@ -92,18 +152,33 @@ export const deleteOrder = (orderDetails) => {
   return async (dispatch) => {
     try {
       const token = window.localStorage.getItem(TOKEN);
-      await axios.delete('/api/orders/current', {
-        headers: {
-          authorization: token,
-        },
-        data: orderDetails,
-      });
-      const { data: newOrder } = await axios.get('/api/orders/current', {
-        headers: {
-          authorization: token,
-        },
-      });
-      dispatch(setCurrentOrder(newOrder));
+      if (token) {
+        await axios.delete('/api/orders/current', {
+          headers: {
+            authorization: token,
+          },
+          data: orderDetails,
+        });
+        const { data: newOrder } = await axios.get('/api/orders/current', {
+          headers: {
+            authorization: token,
+          },
+        });
+        dispatch(setCurrentOrder(newOrder));
+      } else {
+        let cart = JSON.parse(localStorage.getItem("cart"))
+        const arr = cart.products.filter(current => {
+          if (current.id != orderDetails.productId) {
+            return current
+          }
+        })
+        cart = {
+          ...cart,
+          products: arr
+        }
+        localStorage.setItem("cart", JSON.stringify(cart))
+        dispatch(setCurrentOrder(cart))
+      }
     } catch (err) {
       console.error(err);
     }
